Successful implementation of Newton-Raphson

+++ final_project.h
@@ -6,20 +6,20 @@
 #include <string.h>
 #include <ctype.h>
 
-typedef double (*function)(double, double,
+typedef long double (*function)(long double, double,
 			unsigned long **, int,
 				unsigned long **, int);
 int stringNumericCheck(char *str);
 int max(int a, int b);
 int min(int a, int b);
 unsigned long binom(int n, int k);
-double generate_poly(double theta, double alpha,
+long double generate_poly(long double theta, double alpha,
 			unsigned long **numer, int numer_rows,
 				unsigned long **denom, int denom_rows);
-double generate_poly_deriv(double theta, double alpha,
+long double generate_poly_deriv(long double theta, double alpha,
 				unsigned long **numer, int numer_rows,
 					unsigned long **denom, int denom_rows);
-double newton_raphson(function f, function f_prime,
+long double newton_raphson(function f, function f_prime,
 			double theta_0, double alpha, 
 				unsigned long **numer, int numer_rows,
 					unsigned long **denom, int denom_rows);
